pyproject.toml
[project]
name = "trip-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "amadeus>=12.0.0",
    "faster-whisper>=1.1.1",
    "geopy>=2.4.1",
    "ipython>=9.4.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-groq>=0.3.5",
    "langgraph>=0.5.2",
    "livekit-agents[groq,silero]~=1.0",
    "livekit-plugins-groq>=1.2.1",
    "livekit-plugins-silero>=1.2.1",
    "pillow>=11.3.0",
    "pyaudio>=0.2.14",
    "python-dateutil>=2.9.0.post0",
    "scipy>=1.16.0",
    "setuptools<81",
    "sounddevice>=0.5.2",
]
